c03ok

+++ c03/ex03/ft_strncat.c
@@ -1,22 +1,32 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strncat.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mpistaso <mpistaso@student.42barcel>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/12/06 17:03:37 by mpistaso          #+#    #+#             */
+/*   Updated: 2023/12/06 19:21:09 by mpistaso         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 char	*ft_strncat(char *dest, char *src, unsigned int nb)
 {
-int i;
-unsigned int j;
+	int	i;
+	int	j;
 
-i=0;
-j=0;
-
-while (dest[i] != '\0')
-{
-	i++;
-}
-while (src[j] != '\0' && j < nb)
-{
-	dest[i] = src[j];
-	i++;
-	j++;
-}
-dest[i] = '\0';
-
-return(dest);
+	i = 0;
+	j = 0;
+	while (dest[i] != '\0')
+	{
+		i++;
+	}
+	while (src[j] != '\0' && (unsigned int)j < nb)
+	{
+		dest[i] = src[j];
+		i++;
+		j++;
+	}
+	dest[i] = '\0';
+	return (dest);
 }
